requirements.txt
asgiref==3.8.1
base58==2.1.1
Django==4.2.23
django-admin-sortable2==2.2.8
django-user-visit==2.1
gunicorn==23.0.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
whitenoise==6.9.0
